undid circular import hissy fit
