+++ mbuild/formats/hoomdxml.py
@@ -2,10 +2,11 @@ import os
 from warnings import warn
 
 from mdtraj.formats.registry import _FormatRegistry
+import numpy as np
 from six import string_types
 from xml.etree import cElementTree
 
-import mbuild.core
+import mbuild.compound
 
 
 __all__ = ['load_hoomdxml', 'HOOMDTopologyFile']
@@ -64,7 +65,7 @@ class HOOMDTopologyFile(object):
                 raise IOError("The file '%s' doesn't exist" % filename)
             self._fh = open(filename, 'r')
             self._is_open = True
-            self.compound = mbuild.core.Compound()
+            self.compound = mbuild.compound.Compound()
         elif mode == 'w':
             if os.path.exists(filename) and not force_overwrite:
                 raise IOError("The file '%s' already exists" % filename)
@@ -116,7 +117,7 @@ class HOOMDTopologyFile(object):
         atom_mapping = dict()
         for n, type_coord in enumerate(zip(types, coords)):
             atom_type, xyz = type_coord
-            new_atom = mbuild.core.Atom(str(atom_type), [float(x) * self.u['distance']
+            new_atom = mbuild.compound.Atom(str(atom_type), [float(x) * self.u['distance']
                                              for x in xyz.split()])
             self.compound.add(new_atom, label="{0}[$]".format(new_atom.kind))
             atom_mapping[n] = new_atom
@@ -152,18 +153,17 @@ class HOOMDTopologyFile(object):
             try:
                 node_text = self._config.find(node).text.splitlines()[1:]
             except AttributeError:
-                warn("Specified node '{0}' does not exist in {1}".format(
-                    node, self._filename))
+                # Node doesn't exist.
+                pass
             else:
                 if node != 'bond':
                     self.compound.extras[node] = list()
                 for raw_line in node_text:
                     parsed_line = [int(x) if x.isdigit() else x for x in raw_line.split()]
                     if node == 'bond':
-                        kind = atom_mapping[parsed_line[0]]
                         atom1 = atom_mapping[parsed_line[1]]
                         atom2 = atom_mapping[parsed_line[2]]
-                        new_bond = mbuild.core.Bond(atom1, atom2, kind)
+                        new_bond = mbuild.compound.Bond(atom1, atom2)
                         self.compound.add(new_bond)
                     else:
                         self.compound.extras[node] = list()
